+++ changelog.md
@@ -1,3 +1,5 @@
+- Fixed all cases of background flicker when the progress bar is displayed.
+
 v0.6.2
 
 - Fixed an issue where certain modded resources failed to load when the game was running on systems using the Turkish locale (and possibly other locales). Reported by Star☆Child.